+++ rust/pylance/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "lance"
-version = "0.1.3"
+name = "pylance"
+version = "0.1.2"
 edition = "2021"
 authors = ["Lance Devs <dev@eto.ai>"]
-description = "A columnar data format that is 100x faster than Parquet for random access."
+description = "python wrapper for lance-rs"
 license = "Apache-2.0"
 repository = "https://github.com/eto-ai/lance"
 readme = "README.md"
@@ -22,7 +22,15 @@ categories = [
     "science"
 ]
 
+
+[lib]
+# `#[pymodule]` in lib.rs
+name = "lance"
+# "cdylib" is necessary to produce a shared library for Python to import from.
+crate-type = ["cdylib"]
+
 [dependencies]
+lance = { path = ".." }
 bytes = "1.3"
 arrow-arith = "31.0"
 arrow-array = "31.0"
@@ -31,7 +39,6 @@ arrow-cast = "31.0.0"
 arrow-data = "31.0"
 arrow-ord = "31.0"
 arrow-schema = "31.0"
-arrow-select = "31.0"
 async-recursion = "1.0"
 async-trait = "0.1.60"
 byteorder = "1.4.3"
@@ -45,27 +52,8 @@ tokio = { version = "1.23", features = ["rt-multi-thread"] }
 url = "2.3"
 rand = "0.8.3"
 futures = "0.3"
-uuid = { version = "1.2", features = ["v4"] }
-faiss = {version = "0.11.0", features=["gpu"], optional = true }
-path-absolutize = "3.0.14"
+pyo3 = { version = "0.17.3", features = ["extension-module"] }
+arrow = { version = "31.0.0", features = ["pyarrow"] }
 
 [build-dependencies]
 prost-build = "0.11"
-
-[dev-dependencies]
-clap = { version = "4.1.1", features = ["derive"] }
-criterion = { version = "0.4", features = ["async", "async_tokio"] }
-pprof = { version = "0.11", features = ["flamegraph", "criterion"] }
-tempfile = "3.3.0"
-
-[features]
-cli = ["clap"]
-# faiss = ["faiss"]
-
-[[bin]]
-name = "lq"
-required-features = ["cli"]
-
-[[bench]]
-name = "distance"
-harness = false
